Add check option to openchange_provision
This option checks if the provision is possible. I added it because it is needed for Zentyal
server integration and I think it could be useful for more people.
If you do not find it useful, just clsoe the pull request and I will make a script
for Zentyal |server instead