Make emsabp_search return ordered list of MIDs
It was assumed that this method returned the MIDs ordered
like emsabp_ab_container_enum. This made the dcesrv_NspiSeekEntries return
a bad position.

After this change the MIDs are ordered by the SortType in the pStat paremeter.
Currently only SortTypeDisplayName is fully supported, SortTypePhoneticDisplayName
retrun error and other values are unrecognized and default to not sortX.